requirements.txt
Django>=1.3
djangorestframework
pyodbc
South>=0.7.3
django-celery
jsonfield
ipython
pytz
nose
SQLAlchemy